Created MetadataManagedAttribute interface
