+++ dina-base-api/src/main/java/ca/gc/aafc/dina/entity/MetadataManagedAttribute.java
@@ -1,0 +1,12 @@
+package ca.gc.aafc.dina.entity;
+
+import java.util.Map;
+
+public interface MetadataManagedAttribute extends DinaEntity {
+  
+  ManagedAttribute getManagedAttribute();
+
+  Map.Entry<String,String> getMetaData();
+
+  String getAssignedValue();
+}
