+++ dina-base-api/src/main/resources/validation-messages.properties
@@ -1,0 +1,1 @@
+assignedValue.invalid=assignedValue "{0}" is invalid.
